puppeteer 1.15 move ElementHandle to JSHandle
